Get rid of auto-unboxing in comparison

+++ src/com/fsck/k9/helper/DomainNameChecker.java
@@ -124,7 +124,7 @@ public class DomainNameChecker {
                     List<?> altNameEntry = (List<?>)(subjectAltName);
                     if ((altNameEntry != null) && (2 <= altNameEntry.size())) {
                         Integer altNameType = (Integer)(altNameEntry.get(0));
-                        if (altNameType != null && altNameType == ALT_IPA_NAME) {
+                        if (altNameType != null && altNameType.intValue() == ALT_IPA_NAME) {
                             String altName = (String)(altNameEntry.get(1));
                             if (altName != null) {
                                 if (K9.DEBUG) {
